requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
cryptography==41.0.4
ipykernel==6.25.2
ipython==8.16.1
jupyter_client==8.3.1
jupyter_core==5.3.2
lxml==4.9.3
nested-lookup==0.2.25
numpy==1.26.0
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pandas==2.1.1
platformdirs==3.11.0
playwright==1.38.0
ptyprocess==0.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
selenium==4.14.0
urllib3==1.26.17
urllib3-secure-extra==0.1.0
wcwidth==0.2.8
webdriver-manager==3.2.2
